make fbguard exit to the OS after fbserver is ready
Exiting right after forking creates a race condition making it possible
that the service says it is started, but the fbserver process isn't
ready to serve requests, because it hasn't called `listen()` yet.

To reproduce the problem:

  while true; do sudo service firebird restart; echo "select 'ready' from rdb\$database;" | isql localhost:employee; sleep 2.1; done

(the sleep is there to avoid the systemd limitation of 5 service
restarts per 10 seconds).

The patch makes fbserver send SIGUSR1 to the fbguard parent process,
which waits for SIGUSR1 before it returns the control to the OS.

Part of Debian firebird3.0 packages since 2017. First reported on
<https://sourceforge.net/p/firebird/mailman/message/35769037/>